fix flex tests
