+++ editor/package.json
@@ -161,6 +161,7 @@
     "@shopify/hydrogen": "2024.4.1",
     "@stitches/react": "1.2.8",
     "@svgr/plugin-jsx": "5.5.0",
+    "@testing-library/user-event": "14.5.2",
     "@tippyjs/react": "4.1.0",
     "@twind/core": "1.1.3",
     "@twind/preset-autoprefix": "1.0.7",
