+++ editor/src/components/inspector/add-remove-layout-system-control.spec.browser2.tsx
@@ -1,14 +1,10 @@
 import { shiftModifier } from '../../utils/modifiers'
 import { expectSingleUndo2Saves } from '../../utils/utils.test-utils'
-import { ItemContainerTestId } from '../../uuiui/radix-components'
 import { CanvasControlsContainerID } from '../canvas/controls/new-canvas-controls'
 import { mouseClickAtPoint, pressKey } from '../canvas/event-helpers.test-utils'
 import type { EditorRenderResult } from '../canvas/ui-jsx.test-utils'
 import { renderTestEditorWithCode } from '../canvas/ui-jsx.test-utils'
-import {
-  AddFlexLayoutOptionId,
-  AddRemoveLayoutSystemControlTestId,
-} from './add-remove-layout-system-control'
+import { addFlexLayout, removeFlexLayout } from './layout-systems.test-utils'
 
 describe('add layout system', () => {
   it('add and remove layout system via keyboard shortcut', async () => {
@@ -39,13 +35,13 @@ describe('add layout system', () => {
     )
     const div = await selectDiv(editor)
     await expectSingleUndo2Saves(editor, async () => {
-      await clickOn(editor)
+      await addFlexLayout(editor)
     })
 
     expect(div.style.display).toEqual('flex')
 
     await expectSingleUndo2Saves(editor, async () => {
-      await clickOn(editor)
+      await removeFlexLayout(editor)
     })
 
     expect(div.style.display).toEqual('')
@@ -58,7 +54,7 @@ describe('add layout system', () => {
     )
     const div = await selectDiv(editor)
     await expectSingleUndo2Saves(editor, async () => {
-      await clickOn(editor)
+      await addFlexLayout(editor)
     })
 
     expect(div.style.display).toEqual('flex')
@@ -77,7 +73,7 @@ describe('add layout system', () => {
     )
     const div = await selectDiv(editor)
     await expectSingleUndo2Saves(editor, async () => {
-      await clickOn(editor)
+      await addFlexLayout(editor)
     })
 
     expect(div.style.display).toEqual('flex')
@@ -125,7 +121,7 @@ describe('add layout system', () => {
       'await-first-dom-report',
     )
     const div = await selectDiv(editor)
-    await clickOn(editor)
+    await addFlexLayout(editor)
 
     expect(div.style.display).toEqual('flex')
 
@@ -175,7 +171,7 @@ describe('add layout system', () => {
       'await-first-dom-report',
     )
     const div = await selectDiv(editor)
-    await clickOn(editor)
+    await addFlexLayout(editor)
 
     expect(div.style.display).toEqual('flex')
 
@@ -207,15 +203,6 @@ async function selectDiv(editor: EditorRenderResult): Promise<HTMLElement> {
   await mouseClickAtPoint(canvasControlsLayer, divCorner)
 
   return div
-}
-
-async function clickOn(editor: EditorRenderResult) {
-  const flexDirectionToggle = editor.renderedDOM.getAllByTestId(
-    AddRemoveLayoutSystemControlTestId(),
-  )[0]
-  await mouseClickAtPoint(flexDirectionToggle, { x: 2, y: 2 })
-  const flexOption = editor.renderedDOM.getByTestId(ItemContainerTestId(AddFlexLayoutOptionId))
-  await mouseClickAtPoint(flexOption, { x: 2, y: 2 })
 }
 
 function project({ width, height }: { width: string; height: string }): string {
